% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot MDS for a DESeq2 object}
\usage{
plotMDS(rld)
}
\arguments{
\item{rld}{A DESeqTransform object.}
}
\description{
Plot MDS for a DESeq2 object
}
\examples{
library(cowplot)
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
plotMDS(rld)
Cghinga<-tu+madre
}
